requirements.txt
asgiref==3.7.2
Django==5.0.2
django-autoslug==1.9.9
django-cities==0.6.1
django-cities-light==3.10.1
django-countries==7.5.1
django-model-utils==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.3
progressbar2==4.4.2
python-stdnum==1.19
python-utils==3.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
Unidecode==1.3.8
